Update Fluidsynth to v2.0.2

+++ libs/fluidsynth/src/fluid_iir_filter.c
@@ -275,8 +275,8 @@ fluid_iir_filter_calculate_coefficients(fluid_iir_filter_t *iir_filter,
          * into account for both significant frequency relocation and for
          * bandwidth readjustment'. */
 
-        fluid_real_t omega = (fluid_real_t)(2.0 * M_PI *
-                                            (iir_filter->last_fres / ((float) output_rate)));
+        fluid_real_t omega = (fluid_real_t)(2.0 * M_PI) *
+                                            (iir_filter->last_fres / output_rate);
         fluid_real_t sin_coeff = (fluid_real_t) sin(omega);
         fluid_real_t cos_coeff = (fluid_real_t) cos(omega);
         fluid_real_t alpha_coeff = sin_coeff / (2.0f * iir_filter->q_lin);
